move to new config
