+++ lua/codecompanion/adapters/http/anthropic.lua
