+++ lua/codecompanion/adapters/http/azure_openai.lua
@@ -1,4 +1,4 @@
-local openai = require("codecompanion.adapters.openai")
+local openai = require("codecompanion.adapters.http.openai")
 
 ---@class AzureOpenAI.Adapter: CodeCompanion.Adapter
 return {
