+++ lua/codecompanion/adapters/http/deepseek.lua
@@ -1,5 +1,5 @@
 local log = require("codecompanion.utils.log")
-local openai = require("codecompanion.adapters.openai")
+local openai = require("codecompanion.adapters.http.openai")
 local utils = require("codecompanion.utils.adapters")
 
 ---@class DeepSeek.Adapter: CodeCompanion.Adapter
