+++ lua/codecompanion/adapters/http/githubmodels.lua
@@ -1,6 +1,6 @@
 local Job = require("plenary.job")
 local log = require("codecompanion.utils.log")
-local openai = require("codecompanion.adapters.openai")
+local openai = require("codecompanion.adapters.http.openai")
 local utils = require("codecompanion.utils.adapters")
 
 ---@alias GhToken string|nil
