+++ lua/codecompanion/adapters/http/jina.lua
