+++ lua/codecompanion/adapters/http/openai.lua
