+++ lua/codecompanion/config.lua
@@ -11,27 +11,28 @@ local constants = {
 
 local defaults = {
   adapters = {
-    -- HTTP -------------------------------------------------------------------
-    anthropic = "anthropic",
-    azure_openai = "azure_openai",
-    copilot = "copilot",
-    deepseek = "deepseek",
-    gemini = "gemini",
-    githubmodels = "githubmodels",
-    huggingface = "huggingface",
-    novita = "novita",
-    mistral = "mistral",
-    ollama = "ollama",
-    openai = "openai",
-    xai = "xai",
-    jina = "jina",
-    tavily = "tavily",
-    opts = {
-      allow_insecure = false, -- Allow insecure connections?
-      cache_models_for = 1800, -- Cache adapter models for this long (seconds)
-      proxy = nil, -- [protocol://]host[:port] e.g. socks5://127.0.0.1:9999
-      show_defaults = true, -- Show default adapters
-      show_model_choices = true, -- Show model choices when changing adapter
+    http = {
+      anthropic = "anthropic",
+      azure_openai = "azure_openai",
+      copilot = "copilot",
+      deepseek = "deepseek",
+      gemini = "gemini",
+      githubmodels = "githubmodels",
+      huggingface = "huggingface",
+      novita = "novita",
+      mistral = "mistral",
+      ollama = "ollama",
+      openai = "openai",
+      xai = "xai",
+      jina = "jina",
+      tavily = "tavily",
+      opts = {
+        allow_insecure = false, -- Allow insecure connections?
+        cache_models_for = 1800, -- Cache adapter models for this long (seconds)
+        proxy = nil, -- [protocol://]host[:port] e.g. socks5://127.0.0.1:9999
+        show_defaults = true, -- Show default adapters
+        show_model_choices = true, -- Show model choices when changing adapter
+      },
     },
     acp = {
       codex = "codex",
