+++ tests/config.lua
@@ -6,55 +6,57 @@ return {
     SYSTEM_ROLE = "system",
   },
   adapters = {
-    test_adapter = {
-      name = "test_adapter",
-      url = "https://api.openai.com/v1/chat/completions",
-      roles = {
-        llm = "assistant",
-        user = "user",
+    http = {
+      test_adapter = {
+        name = "test_adapter",
+        url = "https://api.openai.com/v1/chat/completions",
+        roles = {
+          llm = "assistant",
+          user = "user",
+        },
+        opts = {
+          stream = true,
+        },
+        headers = {
+          content_type = "application/json",
+        },
+        parameters = {
+          stream = true,
+        },
+        handlers = {
+          form_parameters = function()
+            return {}
+          end,
+          form_messages = function()
+            return {}
+          end,
+          is_complete = function()
+            return false
+          end,
+          tools = {
+            format_tool_calls = function(self, tools)
+              return tools
+            end,
+            output_response = function(self, tool_call, output)
+              return {
+                role = "tool",
+                tool_call_id = tool_call.id,
+                content = output,
+                opts = { tag = tool_call.id, visible = false },
+              }
+            end,
+          },
+        },
+        schema = {
+          model = {
+            default = "gpt-3.5-turbo",
+          },
+        },
       },
       opts = {
-        stream = true,
+        allow_insecure = false,
+        proxy = nil,
       },
-      headers = {
-        content_type = "application/json",
-      },
-      parameters = {
-        stream = true,
-      },
-      handlers = {
-        form_parameters = function()
-          return {}
-        end,
-        form_messages = function()
-          return {}
-        end,
-        is_complete = function()
-          return false
-        end,
-        tools = {
-          format_tool_calls = function(self, tools)
-            return tools
-          end,
-          output_response = function(self, tool_call, output)
-            return {
-              role = "tool",
-              tool_call_id = tool_call.id,
-              content = output,
-              opts = { tag = tool_call.id, visible = false },
-            }
-          end,
-        },
-      },
-      schema = {
-        model = {
-          default = "gpt-3.5-turbo",
-        },
-      },
-    },
-    opts = {
-      allow_insecure = false,
-      proxy = nil,
     },
   },
   strategies = {
